+++ src/Settings/CustomSettings.hpp
@@ -9,14 +9,16 @@ using namespace geode::prelude;
     --- Test ---
 */
 
-class SettingTestValue;
-
-class SettingTestValue : public SettingValue {
-protected:
-    std::string m_placeholder;
+class SettingTestValue : public SettingV3 {
 public:
-    SettingTestValue(std::string const& key, std::string const& modID, std::string const& placeholder)
-      : SettingValue(key, modID), m_placeholder(placeholder) {}
+    static Result<std::shared_ptr<SettingTestValue>> parse(std::string const& key, std::string const& modID, matjson::Value const& json) {
+        auto res = std::make_shared<SettingTestValue>();
+        auto root = checkJson(json, "SettingTestValue");
+        res->init(key, modID, root);
+        res->parseNameAndDescription(root);
+        root.checkUnknownKeys();
+        return root.ok(res);
+    }
 
     bool load(matjson::Value const& json) override {
         return true;
@@ -24,330 +26,50 @@ public:
     bool save(matjson::Value& json) const override {
         return true;
     }
-    SettingNode* createNode(float width) override;
+    bool isDefaultValue() const override {
+        return true;
+    }
+    void reset() override {}
+    SettingNodeV3* createNode(float width) override;
 };
 
 
-class SettingTestNode : public SettingNode {
+class SettingTestNode : public SettingNodeV3 {
 protected:
- bool init(SettingTestValue* value, float width) {
-        if (!SettingNode::init(value))
+    bool init(std::shared_ptr<SettingTestValue> setting, float width) {
+        if (!SettingNodeV3::init(setting, width))
             return false;
         this->setContentSize({ width, 35.f });
-        auto menu = CCMenu::create();
         auto label = CCLabelBMFont::create("Test", "bigFont.fnt");
         label->setScale(.6F);
-        label->setPositionX(-130);
         auto playSpr = CCSprite::createWithSpriteFrameName("GJ_playBtn2_001.png");
         playSpr->setScale(.4F);
-        //auto infoBtn
         auto testBtn = CCMenuItemSpriteExtra::create(
             playSpr,
             this,
             menu_selector(SettingTestNode::onTestBtn)
         );
-        testBtn->setPositionX(140);
-        menu->setPosition(width / 2, 18.f);
-        menu->addChild(label);
-        menu->addChild(testBtn);
-        this->addChild(menu);
+        getButtonMenu()->setContentWidth(width);
+        getButtonMenu()->addChildAtPosition(label, Anchor::Left, {42, 0});
+        getButtonMenu()->addChildAtPosition(testBtn, Anchor::Right, {-15, 0});
+        getButtonMenu()->updateLayout();
+        getNameLabel()->setVisible(false);
         return true;
     }
+    void onCommit() override {}
+    void onResetToDefault() override {}
 
 public:
-    // because for some reason SettingNode doesnt expose description, i have to do all of this, smh.
     void onTestBtn(CCObject*);
-
-    void commit() override {
-        // Let the UI know you have committed the value
-        this->dispatchCommitted();
-    }
-
-    // Geode calls this to query if the setting value has been changed, 
-    // and those changes haven't been committed
-    bool hasUncommittedChanges() override {
+    bool hasUncommittedChanges() const override {
         return false;
     }
-
-    // Geode calls this to query if the setting has a value that is 
-    // different from its default value
-    bool hasNonDefaultValue() override {
-        return true;
-    }
-
-    // Geode calls this to reset the setting's value back to default
-    void resetToDefault() override {
-
-    }
-    static SettingTestNode* create(SettingTestValue* value, float width) {
-        auto ret = new SettingTestNode;
-        if (ret && ret->init(value, width)) {
-            ret->autorelease();
-            return ret;
-        }
-        CC_SAFE_DELETE(ret);
-        return nullptr;
-    }
-};
-
-/*
-    --- Section ---
-*/
-
-class SettingSectionValue;
-
-class SettingSectionValue : public SettingValue {
-protected:
-    std::string m_placeholder;
-public:
-    SettingSectionValue(std::string const& key, std::string const& modID, std::string const& placeholder)
-      : SettingValue(key, modID), m_placeholder(placeholder) {}
-
-    bool load(matjson::Value const& json) override {
-        return true;
-    }
-    bool save(matjson::Value& json) const override {
-        return true;
-    }
-    SettingNode* createNode(float width) override;
-};
-
-class SettingSectionNode : public SettingNode {
-protected:
-    bool init(SettingSectionValue* value, float width) {
-        if (!SettingNode::init(value))
-            return false;
-        this->setContentSize({ width, 40.f });
-        auto menu = CCMenu::create();
-        std::string sectionName = Mod::get()->getSettingDefinition(value->getKey())->get<CustomSetting>()->json->get<std::string>("name");
-        auto infoSpr = CCSprite::createWithSpriteFrameName("GJ_infoIcon_001.png");
-        infoSpr->setScale(.7F);
-        //auto infoBtn
-        auto infoBtn = CCMenuItemSpriteExtra::create(
-            infoSpr,
-            this,
-            menu_selector(SettingSectionNode::onInfoBtn)
-        );
-        infoBtn->setID(value->getKey());
-        infoBtn->setPositionX(Mod::get()->getSettingDefinition(value->getKey())->get<CustomSetting>()->json->get<int>("posX"));
-        auto label = CCLabelBMFont::create(sectionName.c_str(), "goldFont.fnt");
-        
-        label->setScale(.6F);
-        menu->setPosition(width / 2, 23.f);
-        menu->addChild(label);
-        menu->addChild(infoBtn);
-        this->addChild(menu);
-        return true;
-    }
-
-    void onInfoBtn(CCObject* sender) {
-        // i dont want to deal with template hell
-        auto node = reinterpret_cast<CCMenuItemSpriteExtra*>(sender);
-        if (node == nullptr) return;
-        FLAlertLayer::create(nullptr,
-            Mod::get()->getSettingDefinition(node->getID())->get<CustomSetting>()->json->get<std::string>("name").c_str(),
-            Mod::get()->getSettingDefinition(node->getID())->get<CustomSetting>()->json->get<std::string>("description").c_str(),
-            "OK",
-            nullptr,
-            Mod::get()->getSettingDefinition(node->getID())->get<CustomSetting>()->json->get<int>("scale")
-        )->show();
-    }
-
-public:
-    void commit() override {
-        // Let the UI know you have committed the value
-        this->dispatchCommitted();
-    }
-
-    // Geode calls this to query if the setting value has been changed, 
-    // and those changes haven't been committed
-    bool hasUncommittedChanges() override {
+    bool hasNonDefaultValue() const override {
         return false;
     }
-
-    // Geode calls this to query if the setting has a value that is 
-    // different from its default value
-    bool hasNonDefaultValue() override {
-        return true;
-    }
-
-    // Geode calls this to reset the setting's value back to default
-    void resetToDefault() override {
-
-    }
-    static SettingSectionNode* create(SettingSectionValue* value, float width) {
-        auto ret = new SettingSectionNode;
-        if (ret && ret->init(value, width)) {
-            ret->autorelease();
-            return ret;
-        }
-        CC_SAFE_DELETE(ret);
-        return nullptr;
-    }
-};
-
-// Pointercrate / AreDL setting
-
-const int DEFAULT_DL_POS = 2;
-
-struct SettingDLPosStruct {
-    int m_pos;
-};
-
-class SettingDLPosValue;
-
-class SettingDLPosValue : public SettingValue {
-protected:
-    int m_pos;
-public:
-    SettingDLPosValue(std::string const& key, std::string const& modID, int const& position)
-      : SettingValue(key, modID), m_pos(position) {}
-
-    bool load(matjson::Value const& json) override {
-        try {
-            m_pos = static_cast<int>(json.as<int>());
-            return true;
-        } catch(...) {
-            return false;
-        }
-    }
-    bool save(matjson::Value& json) const override {
-        json = static_cast<int>(m_pos);
-        return true;
-    }
-    SettingNode* createNode(float width) override;
-    void setPos(int pos) {
-        m_pos = pos;
-    }
-    int getPos() const {
-        return m_pos;
-    }
-};
-template<>
-struct SettingValueSetter<SettingDLPosStruct> {
-    static SettingDLPosStruct get(SettingValue* setting) {
-        auto posSetting = static_cast<SettingDLPosValue*>(setting);
-        struct SettingDLPosStruct defaultStruct = { posSetting->getPos() };
-        return defaultStruct;
-    };
-    static void set(SettingDLPosValue* setting, SettingDLPosStruct const& value) {
-        setting->setPos(value.m_pos);
-    };
-};
-
-class SettingDLNode : public SettingNode {
-    protected:
-    int m_currentPos;
-    CCMenuItemToggler* pcBtn;
-    CCMenuItemToggler* aredlBtn;
-
-    int getActiveCornerTag(int corner) {
-        switch (corner) {
-            case 1: // AreDL
-                return 2008;
-            case 2: // Pointercrate
-            default:
-                return 2009;
-        }
-    }
-
-    int tagToCorner(int tag) {
-        switch (tag) {
-            case 2008: // AreDL
-                return 1;
-            default:
-            case 2009: // Pointercrate
-                return 2;
-        }
-    }
-
-    bool init(SettingDLPosValue* value, float width) {
-        if (!SettingNode::init(value))
-            return false;
-
-        auto pointercrate_text = CCLabelBMFont::create("Pointercrate", "goldFont.fnt");
-        pointercrate_text->setScale(.6F);
-        pointercrate_text->setPosition(161, 16);
-        this->addChild(pointercrate_text);
-
-        auto aredl_text = CCLabelBMFont::create("AreDL", "goldFont.fnt");
-        aredl_text->setScale(.6F);
-        aredl_text->setPosition(276, 16);
-        this->addChild(aredl_text);
-        
-        m_currentPos = value->getPos();
-        this->setContentSize({ width, 30.f });
-        auto menu = CCMenu::create();
-        CCSprite* toggleOn = CCSprite::createWithSpriteFrameName("GJ_checkOn_001.png");
-        CCSprite* toggleOff = CCSprite::createWithSpriteFrameName("GJ_checkOff_001.png");
-        toggleOn->setScale(.7F);
-        toggleOff->setScale(.7F);
-        menu->setPosition(width / 2, 23.f);
-        pcBtn = CCMenuItemToggler::create(
-            toggleOn,
-            toggleOff,
-            this,
-            menu_selector(SettingDLNode::onCornerClick)
-        );
-        aredlBtn = CCMenuItemToggler::create(
-            toggleOn,
-            toggleOff,
-            this,
-            menu_selector(SettingDLNode::onCornerClick)
-        );
-        pcBtn->setPosition({ -80, -8 });
-        aredlBtn->setPosition({ 60, -8 });
-
-        pcBtn->setTag(getActiveCornerTag(2));
-        aredlBtn->setTag(getActiveCornerTag(1));
-        int currentCorner = m_currentPos;
-        pcBtn->toggle(!(pcBtn->getTag() == getActiveCornerTag(currentCorner)));
-        aredlBtn->toggle(!(aredlBtn->getTag() == getActiveCornerTag(currentCorner)));
-        
-        menu->addChild(pcBtn);
-        menu->addChild(aredlBtn);
-
-        this->addChild(menu);
-        return true;
-    }
-
-    void onCornerClick(CCObject* sender) {
-        pcBtn->toggle(true);
-        aredlBtn->toggle(true);
-        m_currentPos = tagToCorner(sender->getTag());
-        this->dispatchChanged();
-    };
-
-    void onInfoBtn(CCObject* sender) {
-        FLAlertLayer::create(
-            Mod::get()->getSettingDefinition(this->m_value->getKey())->get<CustomSetting>()->json->get<std::string>("name").c_str(),
-            Mod::get()->getSettingDefinition(this->m_value->getKey())->get<CustomSetting>()->json->get<std::string>("description").c_str(),
-            "OK"
-        )->show();
-    }
-    public:
-    void commit() override {
-        static_cast<SettingDLPosValue*>(m_value)->setPos(m_currentPos);
-        this->dispatchCommitted();
-    }
-
-    bool hasUncommittedChanges() override {
-        return m_currentPos != static_cast<SettingDLPosValue*>(m_value)->getPos();
-    }
-
-    bool hasNonDefaultValue() override {
-        return m_currentPos != DEFAULT_DL_POS;
-    }
-
-    void resetToDefault() override {
-        pcBtn->toggle(true);
-        aredlBtn->toggle(false);
-        m_currentPos = DEFAULT_DL_POS;
-    }
-
-    static SettingDLNode* create(SettingDLPosValue* value, float width) {
-        auto ret = new SettingDLNode;
-        if (ret && ret->init(value, width)) {
+    static SettingTestNode* create(std::shared_ptr<SettingTestValue> setting, float width) {
+        auto ret = new SettingTestNode();
+        if (ret && ret->init(setting, width)) {
             ret->autorelease();
             return ret;
         }
@@ -362,97 +84,102 @@ class SettingDLNode : public SettingNode {
 
 const int DEFAULT_POS = 4;
 
-struct SettingPosStruct {
-    int m_pos;
+enum class SettingPosEnum : int {
+    TopLeft,
+    TopRight,
+    BottomLeft,
+    BottomRight
 };
 
-class SettingPosValue;
-
-class SettingPosValue : public SettingValue {
-protected:
-    int m_pos;
-public:
-    SettingPosValue(std::string const& key, std::string const& modID, int const& position)
-      : SettingValue(key, modID), m_pos(position) {}
-
-    bool load(matjson::Value const& json) override {
-        try {
-            m_pos = static_cast<int>(json.as<int>());
-            return true;
-        } catch(...) {
-            return false;
+template <>
+struct matjson::Serialize<SettingPosEnum> {
+    static matjson::Value to_json(SettingPosEnum const& value) {
+        log::info("to_json {}", static_cast<int>(value));
+        switch (value) {
+            case SettingPosEnum::TopLeft:
+                return 1;
+            case SettingPosEnum::TopRight:
+                return 2;
+            case SettingPosEnum::BottomLeft:
+                return 3;
+            default:
+            case SettingPosEnum::BottomRight:
+                return 4;
         }
     }
-    bool save(matjson::Value& json) const override {
-        json = static_cast<int>(m_pos);
-        return true;
+    static SettingPosEnum from_json(matjson::Value const& value) {
+        switch (value.as_int()) {
+            case 1: return SettingPosEnum::TopLeft;
+            case 2: return SettingPosEnum::TopRight;
+            case 3: return SettingPosEnum::BottomLeft;
+            case 4: return SettingPosEnum::BottomRight;
+            default: throw matjson::JsonException(fmt::format("invalid SettingPosEnum value '{}'", value));
+        }
     }
-    SettingNode* createNode(float width) override;
-    void setPos(int pos) {
-        m_pos = pos;
-    }
-    int getPos() const {
-        return m_pos;
+    static bool is_json(matjson::Value const& json) {
+        return json.is_number() || json.is_string();
     }
 };
 
-template<>
-struct SettingValueSetter<SettingPosStruct> {
-    static SettingPosStruct get(SettingValue* setting) {
-        auto posSetting = static_cast<SettingPosValue*>(setting);
-        struct SettingPosStruct defaultStruct = { posSetting->getPos() };
-        return defaultStruct;
-    };
-    static void set(SettingPosValue* setting, SettingPosStruct const& value) {
-        setting->setPos(value.m_pos);
-    };
+class SettingPosValue : public SettingBaseValueV3<SettingPosEnum> {
+public:
+    static Result<std::shared_ptr<SettingPosValue>> parse(std::string const& key, std::string const& modID, matjson::Value const& json) {
+        auto res = std::make_shared<SettingPosValue>();
+        auto root = checkJson(json, "SettingPosValue");
+        res->parseBaseProperties(key, modID, root);
+        root.checkUnknownKeys();
+        return root.ok(res);
+    }
+    SettingNodeV3* createNode(float width) override;
 };
 
-class SettingPosNode : public SettingNode {
+template <>
+struct geode::SettingTypeForValueType<SettingPosEnum> {
+    using SettingType = SettingPosValue;
+};
+
+class SettingPosNode : public SettingValueNodeV3<SettingPosValue> {
 protected:
-    int m_currentPos;
     CCMenuItemToggler* tlBtn;
     CCMenuItemToggler* trBtn;
     CCMenuItemToggler* blBtn;
     CCMenuItemToggler* brBtn;
 
-    int getActiveCornerTag(int corner) {
+    int getActiveCornerTag(SettingPosEnum corner) {
         switch (corner) {
-            case 1: // Top Left
+            case SettingPosEnum::TopLeft: // Top Left
                 return 2004;
-            case 2: // Top Right
+            case SettingPosEnum::TopRight: // Top Right
                 return 2005;
-            case 3: // Bottom Left
+            case SettingPosEnum::BottomLeft: // Bottom Left
                 return 2006;
-            case 4: // Bottom Right
+            case SettingPosEnum::BottomRight: // Bottom Right
             default:
                 return 2007;
         }
     }
-    int tagToCorner(int tag) {
+    SettingPosEnum tagToCorner(int tag) {
         switch (tag) {
             case 2004: // Top Left
-                return 1;
+                return SettingPosEnum::TopLeft;
             case 2005: // Top Right
-                return 2;
+                return SettingPosEnum::TopRight;
             case 2006: // Bottom Left
-                return 3;
+                return SettingPosEnum::BottomLeft;
             default:
             case 2007: // Bottom Right
-                return 4;
+                return SettingPosEnum::BottomRight;
         }
     }
-    bool init(SettingPosValue* value, float width) {
-        if (!SettingNode::init(value))
+
+    bool init(std::shared_ptr<SettingPosValue> setting, float width) {
+        if (!SettingValueNodeV3::init(setting, width))
             return false;
-        m_currentPos = value->getPos();
         this->setContentSize({ width, 70.f });
-        auto menu = CCMenu::create();
         CCSprite* toggleOn = CCSprite::createWithSpriteFrameName("GJ_checkOn_001.png");
         CCSprite* toggleOff = CCSprite::createWithSpriteFrameName("GJ_checkOff_001.png");
         toggleOn->setScale(.7F);
         toggleOff->setScale(.7F);
-        menu->setPosition(width / 2, 23.f);
         tlBtn = CCMenuItemToggler::create(
             toggleOn,
             toggleOff,
@@ -477,242 +204,51 @@ protected:
             this,
             menu_selector(SettingPosNode::onCornerClick)
         );
-        tlBtn->setPosition({ -100, 35 });
-        trBtn->setPosition({ 100, 35 });
-        blBtn->setPosition({ -100, -10 });
-        brBtn->setPosition({ 100, -10 });
-
-        tlBtn->setTag(getActiveCornerTag(1));
-        trBtn->setTag(getActiveCornerTag(2));
-        blBtn->setTag(getActiveCornerTag(3));
-        brBtn->setTag(getActiveCornerTag(4));
-        int currentCorner = m_currentPos;
+        tlBtn->setTag(getActiveCornerTag(SettingPosEnum::TopLeft));
+        trBtn->setTag(getActiveCornerTag(SettingPosEnum::TopRight));
+        blBtn->setTag(getActiveCornerTag(SettingPosEnum::BottomLeft));
+        brBtn->setTag(getActiveCornerTag(SettingPosEnum::BottomRight));
+        getButtonMenu()->setContentSize({width / 2, 45.F});
+        getButtonMenu()->addChildAtPosition(tlBtn, Anchor::TopLeft);
+        getButtonMenu()->addChildAtPosition(trBtn, Anchor::TopRight);
+        getButtonMenu()->addChildAtPosition(blBtn, Anchor::BottomLeft);
+        getButtonMenu()->addChildAtPosition(brBtn, Anchor::BottomRight);
+        getButtonMenu()->updateLayout();
+        getButtonMenu()->setAnchorPoint({0.5, 0.5});
+        getButtonMenu()->updateAnchoredPosition(Anchor::Center, {0, 0});
+        this->getNameLabel()->setVisible(false);
+        auto currentCorner = static_cast<SettingPosEnum>(this->getValue());
         tlBtn->toggle(!(tlBtn->getTag() == getActiveCornerTag(currentCorner)));
         trBtn->toggle(!(trBtn->getTag() == getActiveCornerTag(currentCorner)));
         blBtn->toggle(!(blBtn->getTag() == getActiveCornerTag(currentCorner)));
         brBtn->toggle(!(brBtn->getTag() == getActiveCornerTag(currentCorner)));
-        
-        menu->addChild(tlBtn);
-        menu->addChild(trBtn);
-        menu->addChild(blBtn);
-        menu->addChild(brBtn);
-
-        this->addChild(menu);
+        this->updateState(nullptr);
         return true;
+    }
+    
+    void updateState(CCNode* invoker) override {
+        SettingValueNodeV3::updateState(invoker);
+        auto currentCorner = static_cast<SettingPosEnum>(this->getValue());
     }
     void onCornerClick(CCObject* sender) {
         tlBtn->toggle(true);
         trBtn->toggle(true);
         blBtn->toggle(true);
         brBtn->toggle(true);
-        m_currentPos = tagToCorner(sender->getTag());
-        this->dispatchChanged();
+        SettingPosEnum value = tagToCorner(sender->getTag());
+        this->setValue(value, static_cast<CCNode*>(sender));
     };
-    void onInfoBtn(CCObject* sender) {
-        FLAlertLayer::create(
-            Mod::get()->getSettingDefinition(this->m_value->getKey())->get<CustomSetting>()->json->get<std::string>("name").c_str(),
-            Mod::get()->getSettingDefinition(this->m_value->getKey())->get<CustomSetting>()->json->get<std::string>("description").c_str(),
-            "OK"
-        )->show();
-    }
-public:
-    void commit() override {
-        static_cast<SettingPosValue*>(m_value)->setPos(m_currentPos);
-        this->dispatchCommitted();
-    }
-    bool hasUncommittedChanges() override {
-        return m_currentPos != static_cast<SettingPosValue*>(m_value)->getPos();
-    }
-    bool hasNonDefaultValue() override {
-        return m_currentPos != DEFAULT_POS;
-    }
-
-    // Geode calls this to reset the setting's value back to default
-    void resetToDefault() override {
+    void onResetToDefault() override {
         tlBtn->toggle(true);
         trBtn->toggle(true);
         blBtn->toggle(true);
         brBtn->toggle(false);
-        m_currentPos = DEFAULT_POS;
-    }
-    static SettingPosNode* create(SettingPosValue* value, float width) {
-        auto ret = new SettingPosNode;
-        if (ret && ret->init(value, width)) {
-            ret->autorelease();
-            return ret;
-        }
-        CC_SAFE_DELETE(ret);
-        return nullptr;
-    }
-};
-
-/*
-    Application
-*/
-
-const std::string DEFAULT_APP = "Spotify.exe";
-
-struct SettingAppStruct {
-    std::string m_application;
-};
-
-class SettingAppValue;
-
-class SettingAppValue : public SettingValue {
-protected:
-    std::string m_application;
-public:
-    SettingAppValue(std::string const& key, std::string const& modID, std::string const& application)
-      : SettingValue(key, modID), m_application(application) {}
-
-    bool load(matjson::Value const& json) override {
-        try {
-            m_application = static_cast<std::string>(json.as<std::string>());
-            return true;
-        } catch(...) {
-            return false;
-        }
-    }
-    bool save(matjson::Value& json) const override {
-        json = static_cast<std::string>(m_application);
-        return true;
-    }
-    SettingNode* createNode(float width) override;
-    void setApp(std::string app) {
-        m_application = app;
-    }
-    std::string getApp() const {
-        return m_application;
-    }
-};
-
-template<>
-struct SettingValueSetter<SettingAppStruct> {
-    static SettingAppStruct get(SettingValue* setting) {
-        auto appSetting = static_cast<SettingAppValue*>(setting);
-        struct SettingAppStruct defaultStruct = { appSetting->getApp() };
-        return defaultStruct;
-    };
-    static void set(SettingAppValue* setting, SettingAppStruct const& value) {
-        setting->setApp(value.m_application);
-    };
-};
-
-
-// I LOVE TEMPLATES
-
-class SettingAppNode : public SettingNode {
-protected:
-    std::string m_currentApp;
-    CCMenuItemSpriteExtra* m_resetBtn;
-    bool init(SettingAppValue* value, float width) {
-        if (!SettingNode::init(value))
-            return false;
-        m_currentApp = value->getApp();
-        this->setContentSize({ width, 40.f });
-        auto menu = CCMenu::create();
-        auto label = CCLabelBMFont::create("Application", "bigFont.fnt");
-        label->setScale(.55F);
-        label->setPositionX(-100);
-        auto infoSpr = CCSprite::createWithSpriteFrameName("GJ_infoIcon_001.png");
-        infoSpr->setScale(.6F);
-        
-        //auto infoBtn
-        auto infoBtn = CCMenuItemSpriteExtra::create(
-            infoSpr,
-            this,
-            menu_selector(SettingAppNode::onInfoBtn)
-        );
-        infoBtn->setPositionX(-30);
-        defaultApp_input->setString(m_currentApp);
-        defaultApp_input->setScale(.65F);
-
-        auto fileBtnSpr = CCSprite::createWithSpriteFrameName("gj_folderBtn_001.png");
-        fileBtnSpr->setScale(.5f);
-
-        auto fileBtn = CCMenuItemSpriteExtra::create(
-            fileBtnSpr, this, menu_selector(SettingAppNode::onPickFile)
-        );
-        fileBtn->setPosition(.0f, .0f);
-        fileBtn->setPositionX(143);
-        defaultApp_input->setPosition(70 + (width / 2), 23.f);
-        menu->setPosition(width / 2, 23.f);
-        menu->addChild(label);
-        menu->addChild(infoBtn);
-        this->addChild(defaultApp_input);
-        menu->addChild(fileBtn);
-        this->addChild(menu);
-
-        // Revert button
-        /*auto resetBtnSpr = CCSprite::createWithSpriteFrameName("geode/reset-gold.png");
-        resetBtnSpr->setScale(.5f);
-
-        m_resetBtn = CCMenuItemSpriteExtra::create(
-            resetBtnSpr, this, menu_selector(SettingAppNode::onReset)
-        );
-        m_resetBtn->setPosition(
-            -10,
-            .0f
-        );
-        //if (m_resetBtn) m_resetBtn->setVisible(this->hasNonDefaultValue());
-        menu->addChild(m_resetBtn);*/
-        return true;
-    }
-    void onReset(CCObject*) {
-        createQuickPopup(
-            "Reset",
-            "Are you sure you want to <cr>reset</c> <cl>" + 
-                Mod::get()->getSettingDefinition(this->m_value->getKey())->get<CustomSetting>()->json->get<std::string>("name") +
-                "</c> to <cy>default</c>?",
-            "Cancel", "Reset",
-            [this](auto, bool btn2) {
-                if (btn2) {
-                    this->resetToDefault();
-                }
-            }
-        );
-    }
-    void onPickFile(CCObject*);
-    // nvm cant use this because stupid geode, how are we supposed to get the desc for custom settings??? whats the point of custom settings??
-    void onInfoBtn(CCObject* sender) {
-        /*FLAlertLayer::create(
-            setting()->getDefinition().getDisplayName().c_str(),
-            setting()->castDefinition().description.value(),
-            "OK"
-        )->show();*/
-        FLAlertLayer::create(
-            Mod::get()->getSettingDefinition(this->m_value->getKey())->get<CustomSetting>()->json->get<std::string>("name").c_str(),
-            Mod::get()->getSettingDefinition(this->m_value->getKey())->get<CustomSetting>()->json->get<std::string>("description").c_str(),
-            "OK"
-        )->show();
+        this->setValue(SettingPosEnum::BottomRight, nullptr);
     }
 public:
-    TextInput* defaultApp_input = TextInput::create(180.0F, "Application", "bigFont.fnt");
-    void commit() override {
-        static_cast<SettingAppValue*>(m_value)->setApp(m_currentApp);
-        this->dispatchCommitted();
-    }
-    // Geode calls this to query if the setting value has been changed, 
-    // and those changes haven't been committed
-    bool hasUncommittedChanges() override {
-        return m_currentApp != static_cast<SettingAppValue*>(m_value)->getApp();
-    }
-
-    // Geode calls this to query if the setting has a value that is 
-    // different from its default value
-    bool hasNonDefaultValue() override {
-        return m_currentApp != DEFAULT_APP;
-    }
-
-    // Geode calls this to reset the setting's value back to default
-    void resetToDefault() override {
-        m_currentApp = DEFAULT_APP;
-        static_cast<SettingAppValue*>(m_value)->setApp(m_currentApp);
-        defaultApp_input->setString(DEFAULT_APP);
-    }
-    static SettingAppNode* create(SettingAppValue* value, float width) {
-        auto ret = new SettingAppNode;
-        if (ret && ret->init(value, width)) {
+    static SettingPosNode* create(std::shared_ptr<SettingPosValue> setting, float width) {
+        auto ret = new SettingPosNode();
+        if (ret && ret->init(setting, width)) {
             ret->autorelease();
             return ret;
         }
@@ -725,14 +261,16 @@ public:
     Credits
 */
 
-class SettingCreditsValue;
-
-class SettingCreditsValue : public SettingValue {
-protected:
-    std::string m_placeholder;
+class SettingCreditsValue : public SettingV3 {
 public:
-    SettingCreditsValue(std::string const& key, std::string const& modID, std::string const& placeholder)
-      : SettingValue(key, modID), m_placeholder(placeholder) {}
+    static Result<std::shared_ptr<SettingCreditsValue>> parse(std::string const& key, std::string const& modID, matjson::Value const& json) {
+        auto res = std::make_shared<SettingCreditsValue>();
+        auto root = checkJson(json, "SettingCreditsValue");
+        res->init(key, modID, root);
+        res->parseNameAndDescription(root);
+        root.checkUnknownKeys();
+        return root.ok(res);
+    }
 
     bool load(matjson::Value const& json) override {
         return true;
@@ -740,17 +278,20 @@ public:
     bool save(matjson::Value& json) const override {
         return true;
     }
-    SettingNode* createNode(float width) override;
+    bool isDefaultValue() const override {
+        return true;
+    }
+    void reset() override {}
+    SettingNodeV3* createNode(float width) override;
 };
 
 
-class SettingCreditsNode : public SettingNode {
+class SettingCreditsNode : public SettingNodeV3 {
 protected:
-    bool init(SettingCreditsValue* value, float width) {
-        if (!SettingNode::init(value))
+    bool init(std::shared_ptr<SettingCreditsValue> setting, float width) {
+        if (!SettingNodeV3::init(setting, width))
             return false;
         this->setContentSize({ width, 40.f });
-        auto menu = CCMenu::create();
         auto creditsSpr = ButtonSprite::create("Credits", "bigFont.fnt", "GJ_button_01.png");
         creditsSpr->setScale(.6F);
         auto creditsBtn = CCMenuItemSpriteExtra::create(
@@ -758,7 +299,6 @@ protected:
             this,
             menu_selector(SettingCreditsNode::onCreditsBtn)
         );
-        creditsBtn->setPositionX(-115);
         auto betaTestersSpr = ButtonSprite::create("Beta Testers", "bigFont.fnt", "GJ_button_01.png");
         betaTestersSpr->setScale(.6F);
         auto betaTestersBtn = CCMenuItemSpriteExtra::create(
@@ -766,13 +306,15 @@ protected:
             this,
             menu_selector(SettingCreditsNode::onBetaBtn)
         );
-        betaTestersBtn->setPositionX(90);
-        menu->setPosition(width / 2, 20.f);
-        menu->addChild(creditsBtn);
-        menu->addChild(betaTestersBtn);
-        this->addChild(menu);
+        getButtonMenu()->setContentWidth(width);
+        getButtonMenu()->addChildAtPosition(creditsBtn, Anchor::Center, {-55, 0});
+        getButtonMenu()->addChildAtPosition(betaTestersBtn, Anchor::Center, {85, 0});
+        getButtonMenu()->updateLayout();
+        getNameLabel()->setVisible(false);
         return true;
     }
+    void onCommit() override {}
+    void onResetToDefault() override {}
 
 public:
     void onCreditsBtn(CCObject*) {
@@ -782,31 +324,15 @@ public:
         CreditsBetaMenu::create()->show();
     }
 
-
-    void commit() override {
-        // Let the UI know you have committed the value
-        this->dispatchCommitted();
-    }
-
-    // Geode calls this to query if the setting value has been changed, 
-    // and those changes haven't been committed
-    bool hasUncommittedChanges() override {
+    bool hasUncommittedChanges() const override {
         return false;
     }
-
-    // Geode calls this to query if the setting has a value that is 
-    // different from its default value
-    bool hasNonDefaultValue() override {
-        return true;
+    bool hasNonDefaultValue() const override {
+        return false;
     }
-
-    // Geode calls this to reset the setting's value back to default
-    void resetToDefault() override {
-
-    }
-    static SettingCreditsNode* create(SettingCreditsValue* value, float width) {
-        auto ret = new SettingCreditsNode;
-        if (ret && ret->init(value, width)) {
+    static SettingCreditsNode* create(std::shared_ptr<SettingCreditsValue> setting, float width) {
+        auto ret = new SettingCreditsNode();
+        if (ret && ret->init(setting, width)) {
             ret->autorelease();
             return ret;
         }