+++ changelog.md
@@ -1,4 +1,7 @@
 # Changelog
+## <cj>v1.3.10</c> <cb>(2024-09-15)</c>
+* <cy>Updated</c> settings to the new V3 ones for custom settings.
+
 ## <cj>v1.3.9</c> <cb>(2024-08-19)</c>
 * <cg>Added</c> <cp>Ko-Fi Supporter</c> badge on profile + link to get it on the "gift button" ([or also here](https://ko-fi.com/gdutils))
 * <cg>Added</c> GD moderators leaderboards